+++ test/compojure/api/swagger_test.clj
@@ -3,9 +3,9 @@
             [compojure.api.core :refer :all]
             [compojure.api.swagger :refer :all]
             [compojure.core :refer :all]
-            [midje.sweet :refer :all]
-            [compojure.api.routes :as routes])
-  (:import [java.io StringWriter]))
+            [midje.sweet :refer :all])
+  (:import [java.io StringWriter]
+           [org.joda.time LocalDate]))
 
 (fact "extracting compojure paths"
 
@@ -170,3 +170,19 @@
            '(r1 r2)))
   => {:paths {"/:id"       {:get {:parameters {:path {:id String}}}}
               "/kukka/:id" {:get {:parameters {:path {:id Long}}}}}})
+
+(facts "->path"
+
+  (fact "missing path parameter"
+    (->path "/api/:kikka" {})
+    => (throws IllegalArgumentException))
+
+  (fact "missing serialization"
+    (->path "/api/:kikka" {:kikka (java.security.SecureRandom.)})
+    => (throws com.fasterxml.jackson.core.JsonGenerationException))
+
+  (fact "happy path"
+    (->path "/a/:b/:c/d/:e" {:b (LocalDate/parse "2015-05-22")
+                             :c 12345
+                             :e :kikka})
+    => "/a/2015-05-22/12345/d/kikka"))