+++ examples/src/examples/thingie.clj
@@ -190,5 +190,4 @@
     (POST* "/upload" []
       :multipart-params [file :- TempFileUpload]
       :middlewares [wrap-multipart-params]
-      :consumes ["multipart/form-data"]
       (ok (dissoc file :tempfile)))))